+++ src/Models/Tag.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace c0013r\GhostAPI\Models;
+
+use Carbon\Carbon;
+
+/**
+ * Class Tag
+ * @package c0013r\GhostAPI\Models
+ *
+ * @property-read string $name
+ * @property-read string $description
+ * @property-read string $hidden
+ * @property-read string $parent
+ * @property-read string $featureImage
+ */
+class Tag extends BaseModel
+{
+
+}