+++ src/ServiceProvider.php
@@ -1,0 +1,60 @@
+<?php
+
+namespace c0013r\GhostAPI;
+
+use Illuminate\Support\ServiceProvider as BaseServiceProvider;
+
+class ServiceProvider extends BaseServiceProvider
+{
+    /**
+     * Perform post-registration booting of services.
+     *
+     * @return void
+     */
+    public function boot(): void
+    {
+        // Publishing is only necessary when using the CLI.
+        if ($this->app->runningInConsole())
+        {
+            $this->bootForConsole();
+        }
+    }
+
+    /**
+     * Register any package services.
+     *
+     * @return void
+     */
+    public function register(): void
+    {
+        $this->mergeConfigFrom(__DIR__ . '/../config/ghost.php', 'ghost');
+
+        // Register the service the package provides.
+        $this->app->singleton('Ghost', function ($app) {
+            return new Ghost();
+        });
+    }
+
+    /**
+     * Get the services provided by the provider.
+     *
+     * @return array
+     */
+    public function provides(): array
+    {
+        return ['Ghost'];
+    }
+    
+    /**
+     * Console-specific booting.
+     *
+     * @return void
+     */
+    protected function bootForConsole(): void
+    {
+        // Publishing the configuration file.
+        $this->publishes([
+            __DIR__ . '/../config/ghost.php' => config_path('ghost.php'),
+        ], 'laravel-ghost-api.config');
+    }
+}