+++ src/laravel-ghost-api.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace c0013r\laravel-ghost-api;
-
-class laravel-ghost-api
-{
-    // Build wonderful things
-}