+++ changelog.md
@@ -5,4 +5,4 @@ All notable changes to `laravel-ghost-api` will be documented in this file.
 ## Version 1.0
 
 ### Added
-- Everything
+- Base functions of ghost.org API